+++ src/main/java/controller/ControllerAdapter.java
@@ -1,13 +1,12 @@
-package webserver.controller;
+package controller;
 
-import webserver.ModelAndView;
-import webserver.http.request.HttpRequest;
+import http.request.HttpRequest;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
-public class HandlerAdapter {
+public class ControllerAdapter {
     public static ModelAndView runHandlerMethod(Method method, HttpRequest httpRequest) throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
         Class<?> clazz = method.getDeclaringClass();
         Constructor<?> constructor = clazz.getConstructor();
