+++ src/main/java/http/MIME.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package http;
 
 import utils.Parser;
