+++ src/main/java/http/StatusCode.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package http;
 
 public enum StatusCode {
