+++ src/main/java/http/request/HttpRequest.java
@@ -1,9 +1,9 @@
-package webserver.http.request;
+package http.request;
 
+import http.HttpHeaders;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.Parser;
-import webserver.http.HttpHeaders;
 
 import java.io.BufferedReader;
 import java.io.IOException;
