+++ src/main/java/http/request/HttpRequestLine.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package http.request;
 
 import utils.Parser;
