+++ src/main/java/http/request/RequestMethod.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package http.request;
 
 public enum RequestMethod {
     GET,
