+++ src/main/java/session/UserSessionManager.java
@@ -3,7 +3,7 @@ package session;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.http.Cookie;
+import http.Cookie;
 
 import java.util.Map;
 import java.util.UUID;
