+++ src/main/java/utils/Parser.java
@@ -1,6 +1,6 @@
 package utils;
 
-import webserver.http.Path;
+import http.Path;
 
 import java.io.BufferedReader;
 import java.io.IOException;
