+++ src/main/java/webserver/RequestHandler.java
@@ -2,8 +2,8 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.controller.FrontController;
-import webserver.http.request.HttpRequest;
+import controller.FrontController;
+import http.request.HttpRequest;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
